Combined GlobalMethodLibrary.java and Utility.
